# WIP: Replace unstructured conversion with simple json-iterator conversion