json-iterator to data and then decoding is faster than the current
implementation. This currently uses Fastest to decode but has several unit
test failures which need to be investigated.

@liggitt

/kind bug


The unstructured benchmarks incorrectly charged setup of objects to the benchmark. While doing that I noticed how bad unstructured was now (Go has substantially improved since 1.9 in this area).

Before:

``` 
BenchmarkToUnstructured-12                 	   10000	194565 ns/op	   57428 B/op	     906 allocs/op
BenchmarkFromUnstructured-12               	   10000	    205167 ns/op	   47491 B/op	     796 allocs/op
BenchmarkToUnstructuredViaJSON-12          	   30000	     51490 ns/op	   22610 B/op	     323 allocs/op
BenchmarkFromUnstructuredViaJSON-12        	   10000	    118713 ns/op	   29602 B/op	     438 allocs/op
BenchmarkToUnstructuredViaJSONIter-12      	   50000	     31548 ns/op	   15940 B/op	     390 allocs/op
BenchmarkFromUnstructuredViaJSONIter-12    	   30000	     42272 ns/op	   13468 B/op	     291 allocs/op
```

After:




```release-note
Improve the performance of some conversion significantly.
```