@liggitt 2019-02-12
with the fixed up benchmark tests that correctly assign allocs, I'm seeing this:

before:
```
BenchmarkToUnstructured-8                	   30000	     59201 ns/op	   14110 B/op	     221 allocs/op
BenchmarkFromUnstructured-8              	   30000	     52846 ns/op	    6435 B/op	     185 allocs/op
```

after:
```
BenchmarkToUnstructured-8                	   20000	     53919 ns/op	   20193 B/op	     439 allocs/op
BenchmarkFromUnstructured-8              	   30000	     42873 ns/op	   12310 B/op	     245 allocs/op
```

run time is better, but allocs increased 40-90%